Fix async calls to console.log.
Instead of monkey patching `console.log`, we create a console wrapper
and monkey patch that, passing it as an argument to the generated
function so it will always be in scope.

We keep track of the current console wrapper so that any async calls
from previous evaluations will no longer cause output.

Closes #15.