Update cli.md (#959)
Every (rare) time I need to ignore .babelrc and add some presets/plugins for a custom CLI build I come to this page and it's not clear at all how to do it so I have to go trawling Stack Overflow to find out. I've added examples for these options.