Fix opening of the search input box
Prevent typing of '/' or '?' when either the search box is newly opened
and focused or the box is already opened and the focus is regained.

Additionally don't click the button for opening/closing the box but set
the `clicked`-prop on the `$toggleCheckbox` element, so that it is
definitely just opened.