Rename "Discourse" link to "Forum"
People don't really care that Babel is using Discourse for the forum, just that it actually **is** a forum. Discourse is just an implementation detail. You say "send me an email" not "send me a Gmail", and "give me a phone call" not "contact me via the AT&T network", so this link should read "Forum" or "Discuss" rather than "Discourse".

@hzoo

People that don't have experience with Discourse would be confused by the current link.