Blog: Removing Babel's Stage Presets (#1735)
The purpose of this is to explain the decision to remove the Stage presets but also to give context into why and the history of not just TC39/Babel but the community and it's relationship with the process if that makes sense and where we are at today. Ideally we'd have some more takeaways regarding getting involved in the project.