Update new default value
This bit me while upgrading to babel 7 w/ present-env.

It says the default is commonjs, however that's not true. When I turn debug on and I don't specify a modules value it says it defaults to auto.

This adds auto as an option and denotes it being the default value.