Update "Babel's Funding Plans" blog post to include my sponsored employment
I don't think this was intentional but it was disappointing to see the financial contributions of the companies I've worked for ignored, which I would consider fairly substantial since Babel would be nothing without it. It represents what was then a sustainable model for my work on Babel.

Happy to make changes, and discuss whether I am even considered a "maintainer" but the project and funding did not start when Henry quit his job.