Update usage.md (#2479)
* Update usage.md

Updated usage.md to not recommend the use of `[@babel/polyfill](https://babeljs.io/docs/en/babel-polyfill)`, which is deprecated. Solves #2171

* Update usage.md to include usage of core-js polyfill

Update the documentation to state first how to poliyfill with core-js. Solves #2240, #2163, #2332 and #2006

* fix typo in previous commit

Fix in usage.md where I was referring(linking) to where deprecated was stated

* update usage.md to describe regenerator-runtime polyfills

I updated this doc to describe the use of regenerator-runtime polyfills and when to use them.