Update polyfill.md
After my testing, it depends on the version of the @ babel/plugin transform generator plugin whether the auxiliary function for the generator will be produced. It is not related to the version of @ babel/core.

If the transformer generator plugin version is lower than 7.18.0, the compiled file will not inline the generator auxiliary function. If the version of the transform regenerator plug-in is higher than or equal to 7.18.0, auxiliary inline function will be generated.

Here is a warehouse address for my testing: https://github.com/keer-tea/babel-regenerator-test

There is one type of warehouse, with versions of both packages less than 7.18.0. In Warehouse 2, the @ babel/core version is less than 7.18.0, and the @ babel/plugin transform generator version is greater than 7.18.0. In Warehouse 3, the @ babel/core version is greater than 7.18.0, and the @ babel/plugin transform generator version is less than 7.18.0.

I hope to correct this mistake.